requirements.txt
# requirements.txt
torch>=1.9.0
tqdm>=4.62.0
numpy>=1.20.0
pillow>=8.2.0
mahotas>=1.4.12
matplotlib>=3.4.0
torchvision>=0.10.0
opencv-python>=4.5.0
scikit-image>=0.18.0